fix(evidence): trust transport boundary for replay gate (#620)
