+++ internal/message/message.go
@@ -705,7 +705,7 @@ func DeliverMessage(postPath string, contextID string, knownNodes map[string]dis
 			policyInput.EvidencePresenceGateChecked = true
 			observedAt := evidenceGateObservedAt(sourceSessionDir, sourceSessionName, filename, postPath, time.Now().UTC())
 			policyInput.EvidencePresenceGateActive = cfg.EvidencePresenceGateActiveAt(observedAt)
-			senderBody, senderBodyExact := envelope.SenderBodyFromContent(messageContent)
+			senderBody, senderBodyExact := envelope.SenderBodyFromTrustedContent(messageContent, filename)
 			policyInput.CompletionClaim = senderBodyExact && isCompletionClaim(senderBody)
 			policyInput.EvidencePresent = hasEvidenceReplayContract(metadata)
 			if decision := planDeliveryPolicy(policyInput); decision.Action == deliveryActionDeadLetter {
